v6.35 release

+++ system/SetupSystem.cpp
@@ -1176,8 +1176,8 @@ void PrintHexBytes(const void * vbuf, uint32 numBytes, const char * optDesc, uin
       fprintf(optFile, "%s", headBuf);
 
       const int hexBufSize = (numColumns*8)+1;
-      size_t numDashes = 8+(4*numColumns)-strlen(headBuf);
-      for (size_t i=0; i<numDashes; i++) fputc('-', optFile);
+      int numDashes = 8+(4*numColumns)-(int)strlen(headBuf);
+      for (int i=0; i<numDashes; i++) fputc('-', optFile);
       fputc('\n', optFile);
       if (buf)
       {
@@ -1240,8 +1240,8 @@ void PrintHexBytes(const Queue<uint8> & buf, const char * optDesc, uint32 numCol
       fprintf(optFile, "%s", headBuf);
 
       const int hexBufSize = (numColumns*8)+1;
-      size_t numDashes = 8+(4*numColumns)-strlen(headBuf);
-      for (size_t i=0; i<numDashes; i++) fputc('-', optFile);
+      int numDashes = 8+(4*numColumns)-(int)strlen(headBuf);
+      for (int i=0; i<numDashes; i++) fputc('-', optFile);
       fputc('\n', optFile);
       char * ascBuf = newnothrow_array(char, numColumns+1);
       char * hexBuf = newnothrow_array(char, hexBufSize);
@@ -1290,8 +1290,8 @@ void LogHexBytes(int logLevel, const void * vbuf, uint32 numBytes, const char * 
       LogTime(logLevel, "%s", headBuf);
 
       const int hexBufSize = (numColumns*8)+1;
-      size_t numDashes = 8+(4*numColumns)-strlen(headBuf);
-      for (size_t i=0; i<numDashes; i++) Log(logLevel, "-");
+      int numDashes = 8+(4*numColumns)-(int)strlen(headBuf);
+      for (int i=0; i<numDashes; i++) Log(logLevel, "-");
       Log(logLevel, "\n");
       if (buf)
       {
@@ -1342,8 +1342,8 @@ void LogHexBytes(int logLevel, const Queue<uint8> & buf, const char * optDesc, u
       Log(logLevel, "%s", headBuf);
 
       const int hexBufSize = (numColumns*8)+1;
-      size_t numDashes = 8+(4*numColumns)-strlen(headBuf);
-      for (size_t i=0; i<numDashes; i++) Log(logLevel, "-");
+      int numDashes = 8+(4*numColumns)-(int)strlen(headBuf);
+      for (int i=0; i<numDashes; i++) Log(logLevel, "-");
       Log(logLevel, "\n");
       char * ascBuf = newnothrow_array(char, numColumns+1);
       char * hexBuf = newnothrow_array(char, hexBufSize);
@@ -1403,8 +1403,8 @@ String HexBytesToAnnotatedString(const void * vbuf, uint32 numBytes, const char 
       ret += headBuf;
 
       const int hexBufSize = (numColumns*8)+1;
-      size_t numDashes = 8+(4*numColumns)-strlen(headBuf);
-      for (size_t i=0; i<numDashes; i++) ret += '-';
+      int numDashes = 8+(4*numColumns)-(int)strlen(headBuf);
+      for (int i=0; i<numDashes; i++) ret += '-';
       ret += '\n';
       if (buf)
       {
@@ -1458,8 +1458,8 @@ String HexBytesToAnnotatedString(const Queue<uint8> & buf, const char * optDesc,
       ret += headBuf;
 
       const int hexBufSize = (numColumns*8)+1;
-      size_t numDashes = 8+(4*numColumns)-strlen(headBuf);
-      for (size_t i=0; i<numDashes; i++) ret += '-';
+      int numDashes = 8+(4*numColumns)-(int)strlen(headBuf);
+      for (int i=0; i<numDashes; i++) ret += '-';
       ret += '\n';
       char * ascBuf = newnothrow_array(char, numColumns+1);
       char * hexBuf = newnothrow_array(char, hexBufSize);
